feat: Add subscription model for user subscriptions
This commit adds the `subscription.model.js` file to the `models` directory. The file defines the schema for user subscriptions, including the subscriber and channel fields. The subscriber field references the user who is subscribing, while the channel field references the user to whom the subscriber is subscribing. This model will be used to manage user subscriptions in the application.